requirements.txt
amqp==2.5.2
asgiref==3.2.7
billiard==3.6.3.0
celery==4.4.2
certifi==2020.4.5.1
chardet==3.0.4
cloudipsp==1.0.0
coreapi==2.3.3
coreschema==0.0.4
distro==1.5.0
Django==3.0.5
django-rest-swagger==2.2.0
djangorestframework==3.11.0
flower==0.9.4
humanize==0.5.1
idna==2.9
importlib-metadata==1.6.0
itypes==1.2.0
Jinja2==2.11.2
kombu==4.6.8
MarkupSafe==1.1.1
openapi-codec==1.3.2
Pillow==7.1.1
PyJWT==1.7.1
pytz==2019.3
redis==3.4.1
requests==2.23.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.1
tornado==6.0.4
twilio==6.38.1
uritemplate==3.0.1
urllib3==1.25.9
uuid==1.30
vine==1.3.0
yandex-checkout==1.6.1
zipp==3.1.0
